refactor(gui): move feature-request button into TitleBar
Pre-TitleBar, featureBar was a QToolBar workaround created solely to
host the feature-request icon — macOS's native menu bar couldn't carry
a custom corner widget. Now that TitleBar hosts menus in-window on
every platform, featureBar is redundant dead weight.

Consolidate: TitleBar owns the button; emits featureRequestClicked()
when pressed. MainWindow connects that to its existing
showFeatureRequestDialog slot (unchanged). Delete the 70-line
featureBar block in MainWindow.cpp and the stale m_featureBtn member
in MainWindow.h. Button sits at the far right of the TitleBar strip,
past MasterOutputWidget — matches AetherSDR's pattern where the
feature button is the rightmost element.
