+++ src/gui/MainWindow.cpp
@@ -313,7 +313,6 @@ warren@wpratt.com
 #include <QTimer>
 #include <QThread>
 #include <QPushButton>
-#include <QToolBar>
 #include <QClipboard>
 #include <QDesktopServices>
 #include <QUrl>
@@ -364,6 +363,12 @@ MainWindow::MainWindow(QWidget* parent)
             engine->setSpeakersConfig(cfg);
         }
     });
+
+    // Phase 3O Sub-Phase 10 Task 10d — the 💡 feature-request button
+    // now lives inside TitleBar (consolidated from the old featureBar
+    // QToolBar). Wire its click signal to the existing slot.
+    connect(m_titleBar, &TitleBar::featureRequestClicked,
+            this, &MainWindow::showFeatureRequestDialog);
 
     buildStatusBar();
     applyDarkTheme();
@@ -1592,83 +1597,6 @@ void MainWindow::buildMenuBar()
         AboutDialog dlg(this);
         dlg.exec();
     });
-
-    // =========================================================================
-    // 💡 ISSUE REPORTER — toolbar button (Phase 3G-14)
-    // Ported from AetherSDR TitleBar::showFeatureRequestDialog()
-    // Uses a QToolBar so the button is visible on macOS (native menu bar)
-    // and on Linux/Windows (in-window menu bar) alike.
-    // =========================================================================
-    auto* featureBar = new QToolBar(this);
-    featureBar->setMovable(false);
-    featureBar->setFloatable(false);
-    featureBar->setIconSize(QSize(28, 28));
-    featureBar->setFixedHeight(32);
-    featureBar->setStyleSheet(QStringLiteral(
-        "QToolBar { background: #0f0f1a; border-bottom: 1px solid #203040; spacing: 0; padding: 0 4px; }"));
-
-    auto* spacer = new QWidget;
-    spacer->setSizePolicy(QSizePolicy::Expanding, QSizePolicy::Preferred);
-    featureBar->addWidget(spacer);
-
-    // Paint a lightbulb icon so it renders cleanly at any DPI
-    auto makeBulbIcon = [](QColor bulbColor, QColor baseColor) -> QIcon {
-        constexpr int sz = 64;  // paint large, Qt scales down
-        QPixmap pm(sz, sz);
-        pm.fill(Qt::transparent);
-        QPainter p(&pm);
-        p.setRenderHint(QPainter::Antialiasing);
-
-        // Bulb (circle)
-        p.setPen(Qt::NoPen);
-        p.setBrush(bulbColor);
-        p.drawEllipse(QRectF(14, 4, 36, 36));
-
-        // Neck (trapezoid connecting bulb to base)
-        QPolygonF neck;
-        neck << QPointF(22, 36) << QPointF(42, 36)
-             << QPointF(40, 44) << QPointF(24, 44);
-        p.drawPolygon(neck);
-
-        // Base (screw threads — 3 thin lines)
-        p.setPen(QPen(baseColor, 2.5));
-        p.drawLine(QPointF(24, 46), QPointF(40, 46));
-        p.drawLine(QPointF(25, 50), QPointF(39, 50));
-        p.drawLine(QPointF(27, 54), QPointF(37, 54));
-
-        // Tip
-        p.setPen(Qt::NoPen);
-        p.setBrush(baseColor);
-        p.drawEllipse(QRectF(29, 56, 6, 4));
-
-        // Filament lines inside bulb
-        p.setPen(QPen(baseColor, 1.5));
-        p.drawLine(QPointF(28, 34), QPointF(28, 22));
-        p.drawLine(QPointF(28, 22), QPointF(32, 16));
-        p.drawLine(QPointF(32, 16), QPointF(36, 22));
-        p.drawLine(QPointF(36, 22), QPointF(36, 34));
-
-        p.end();
-        return QIcon(pm);
-    };
-
-    QIcon bulbIcon = makeBulbIcon(QColor(0xFF, 0xD0, 0x60), QColor(0x80, 0x60, 0x20));
-
-    m_featureBtn = new QPushButton;
-    m_featureBtn->setIcon(bulbIcon);
-    m_featureBtn->setIconSize(QSize(22, 22));
-    m_featureBtn->setFixedSize(28, 28);
-    m_featureBtn->setToolTip(QStringLiteral("Submit a feature request or bug report"));
-    m_featureBtn->setAccessibleName(QStringLiteral("Feature request"));
-    m_featureBtn->setStyleSheet(QStringLiteral(
-        "QPushButton { background: #3a2a00; border: 1px solid #806020; "
-        "border-radius: 4px; padding: 0; }"
-        "QPushButton:hover { background: #504000; border-color: #a08030; }"));
-    connect(m_featureBtn, &QPushButton::clicked,
-            this, &MainWindow::showFeatureRequestDialog);
-    featureBar->addWidget(m_featureBtn);
-
-    addToolBar(Qt::TopToolBarArea, featureBar);
 }
 
 void MainWindow::buildStatusBar()
