+++ src/gui/MainWindow.h
@@ -75,7 +75,6 @@
 #include <QActionGroup>
 #include <QTimer>
 
-class QPushButton;
 class QProgressDialog;
 class QThread;
 class QSplitter;
@@ -216,9 +215,6 @@ private:
 
     // Spectrum overlay panel
     class SpectrumOverlayPanel* m_overlayPanel{nullptr};
-
-    // Phase 3G-14: Issue reporter corner widget
-    QPushButton* m_featureBtn{nullptr};
 
     // Applet panel — scrollable content widget inside Container #0
     class AppletPanelWidget* m_appletPanel{nullptr};
